+++ ibis/backends/base/df/scope.py
@@ -37,13 +37,14 @@ different time contexts.
 from __future__ import annotations
 
 from collections import namedtuple
-from typing import TYPE_CHECKING, Any, Iterable
+from typing import Any, Iterable, Tuple
 
+import pandas as pd
+
+from ibis.backends.base.df.timecontext import TimeContextRelation, compare_timecontext
 from ibis.expr.operations import Node
-from ibis.expr.timecontext import TimeContextRelation, compare_timecontext
 
-if TYPE_CHECKING:
-    from ibis.expr.typing import TimeContext
+TimeContext = Tuple[pd.Timestamp, pd.Timestamp]
 
 ScopeItem = namedtuple('ScopeItem', ['timecontext', 'value'])
