+++ ibis/backends/dask/tests/execution/test_timecontext.py
@@ -1,23 +1,19 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 import pytest
 from pandas import Timedelta, Timestamp
 
 import ibis
 import ibis.common.exceptions as com
 import ibis.expr.operations as ops
-from ibis.backends.pandas.execution import execute
-from ibis.expr.scope import Scope
-from ibis.expr.timecontext import (
+from ibis.backends.base.df.scope import Scope
+from ibis.backends.base.df.timecontext import (
+    TimeContext,
     TimeContextRelation,
     adjust_context,
     compare_timecontext,
 )
-
-if TYPE_CHECKING:
-    from ibis.expr.typing import TimeContext
+from ibis.backends.pandas.execution import execute
 
 dd = pytest.importorskip("dask.dataframe")
 from dask.dataframe.utils import tm  # noqa: E402
