+++ ibis/backends/pandas/dispatch.py
@@ -7,8 +7,8 @@ from multipledispatch import Dispatcher
 import ibis.common.exceptions as com
 import ibis.expr.operations as ops
 from ibis.backends.base import BaseBackend
+from ibis.backends.base.df.scope import Scope
 from ibis.backends.pandas.trace import TraceTwoLevelDispatcher
-from ibis.expr.scope import Scope
 
 # Individual operation execution
 execute_node = TraceTwoLevelDispatcher(
