+++ ibis/backends/pandas/execution/temporal.py
@@ -9,6 +9,7 @@ from pandas.core.groupby import SeriesGroupBy
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 from ibis.backends.base import BaseBackend
+from ibis.backends.base.df.scope import Scope
 from ibis.backends.pandas.core import (
     date_types,
     integer_types,
@@ -18,7 +19,6 @@ from ibis.backends.pandas.core import (
 )
 from ibis.backends.pandas.dispatch import execute_node, pre_execute
 from ibis.backends.pandas.execution.util import get_grouping
-from ibis.expr.scope import Scope
 
 
 @execute_node.register(ops.Strftime, pd.Timestamp, str)
