+++ ibis/backends/pandas/tests/test_core.py
@@ -8,11 +8,11 @@ import ibis
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
+from ibis.backends.base.df.scope import Scope
 from ibis.backends.pandas import Backend
 from ibis.backends.pandas.core import is_computable_input
 from ibis.backends.pandas.dispatch import execute_node, post_execute, pre_execute
 from ibis.backends.pandas.execution import execute
-from ibis.expr.scope import Scope
 
 
 @pytest.fixture
