+++ ibis/backends/pyspark/timecontext.py
@@ -1,15 +1,10 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 import pyspark.sql.functions as F
 from pyspark.sql.dataframe import DataFrame
 
 import ibis.common.exceptions as com
-from ibis.expr.timecontext import get_time_col
-
-if TYPE_CHECKING:
-    from ibis.expr.typing import TimeContext
+from ibis.backends.base.df.timecontext import TimeContext, get_time_col
 
 
 def filter_by_time_context(
